ErasedTagIndex to tag_index crate
