+++ src/redisearch_rs/tag_index/tests/integration/main.rs
@@ -18,6 +18,7 @@ redis_mock::mock_or_stub_missing_redis_c_symbols!();
 mod create;
 mod disk;
 mod disk_backend;
+mod erased;
 mod expansion_timeout;
 mod filtered_iteration;
 mod gc;
